refactor(checkbox): interfaces
