+++ ui-parts/checkbox/src/index.ts
@@ -5,9 +5,5 @@ export type * from './checkbox.interfaces.js'
 export * from './container/index.js'
 export * from './hidden-input/index.js'
 export * from './label/index.js'
-export type { CheckboxAppearance } from './styles/index.js'
-export type { CheckboxAppearanceName } from './styles/index.js'
-export type { CheckboxShape } from './styles/index.js'
-export type { CheckboxShapeName } from './styles/index.js'
 export { checkboxAppearances } from './styles/index.js'
 export { checkboxShapes } from './styles/index.js'
