+++ ui-parts/checkbox/src/styles/index.ts
@@ -1,3 +1,2 @@
 export * from './appearance.css.js'
-export type * from './interfaces.js'
 export * from './shape.css.js'
