--- ui-parts/checkbox/src/styles/interfaces.ts
@@ -1,15 +1,0 @@
-export type CheckboxAppearanceName = 'blue' | 'green' | 'red'
-
-export interface CheckboxAppearance {
-  box: string
-  check: string
-  label: string
-}
-
-export type CheckboxShapeName = 'large' | 'medium' | 'small'
-
-export interface CheckboxShape {
-  box: string
-  check: string
-  label: string
-}
